+++ android/KMAPro/kMAPro/src/main/java/com/tavultesoft/kmapro/KeymanSettingsFragment.java
@@ -3,7 +3,10 @@ package com.tavultesoft.kmapro;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.provider.Settings;
+import android.view.inputmethod.InputMethodManager;
 
+import androidx.preference.CheckBoxPreference;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
 import androidx.preference.PreferenceScreen;
@@ -12,14 +15,13 @@ import androidx.preference.SwitchPreference;
 import com.tavultesoft.kmea.KMManager;
 import com.tavultesoft.kmea.data.Dataset;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-
 public class KeymanSettingsFragment extends PreferenceFragmentCompat {
   private static final String TAG = "SettingsFragment";
   private static Context context;
 
   private Preference languagesPreference;
+  private CheckBoxPreference setSystemKeyboardPreference;
+  private CheckBoxPreference setDefaultKeyboardPreference;
 
   @Override
   public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
@@ -35,7 +37,6 @@ public class KeymanSettingsFragment extends PreferenceFragmentCompat {
     languagesPreference.setWidgetLayoutResource(R.layout.preference_icon_layout);
     Intent languagesIntent = new Intent();
     languagesIntent.setClassName(context.getPackageName(), "com.tavultesoft.kmea.LanguagesSettingsActivity");
-    languagesIntent.addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY);
     languagesIntent.putExtra(KMManager.KMKey_DisplayKeyboardSwitcher, false);
     languagesPreference.setIntent(languagesIntent);
 
@@ -56,7 +57,44 @@ public class KeymanSettingsFragment extends PreferenceFragmentCompat {
     getStartedPreference.setTitle(getString(R.string.show_get_started));
     getStartedPreference.setDefaultValue(true);
 
+    // Blocks the default checkmark interaction; we want to control the checkmark's state separately
+    // from within update() based on if the user has taken the appropriate actions with the OS.
+    final Preference.OnPreferenceChangeListener checkBlocker = new Preference.OnPreferenceChangeListener() {
+      @Override
+      public boolean onPreferenceChange(Preference preference, Object newValue) {
+        return false;
+      }
+    };
+
+    setSystemKeyboardPreference = new CheckBoxPreference(context);
+    setSystemKeyboardPreference.setTitle(R.string.enable_system_keyboard);
+    setSystemKeyboardPreference.setDefaultValue(SystemIMESettings.isEnabledAsSystemKB(context));
+    setSystemKeyboardPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+      @Override
+      public boolean onPreferenceClick(Preference preference) {
+        startActivity(new Intent(Settings.ACTION_INPUT_METHOD_SETTINGS));
+        return false;
+      }
+    });
+    setSystemKeyboardPreference.setOnPreferenceChangeListener(checkBlocker);
+
+    setDefaultKeyboardPreference = new CheckBoxPreference(context);
+    setDefaultKeyboardPreference.setTitle(R.string.set_keyman_as_default);
+    setDefaultKeyboardPreference.setDefaultValue(SystemIMESettings.isDefaultKB(context));
+    setDefaultKeyboardPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+      @Override
+      public boolean onPreferenceClick(Preference preference) {
+        InputMethodManager imManager = (InputMethodManager) context.getSystemService(Context.INPUT_METHOD_SERVICE);
+        imManager.showInputMethodPicker();
+        return false;
+      }
+    });
+    setDefaultKeyboardPreference.setOnPreferenceChangeListener(checkBlocker);
+
     screen.addPreference(languagesPreference);
+    screen.addPreference(setSystemKeyboardPreference);
+    screen.addPreference(setDefaultKeyboardPreference);
+
     screen.addPreference(bannerPreference);
     screen.addPreference(getStartedPreference);
 
@@ -86,6 +124,17 @@ public class KeymanSettingsFragment extends PreferenceFragmentCompat {
   @Override
   public void onResume() {
     super.onResume();
+
+    // The onResume() function isn't called after the default-keyboard selection and cannot fix that
+    // option, though it is sufficient for handling the system-keyboard enablement option.
+    //
+    // As a result, we rely on KeymanSettingsActivity.onWindowFocusChanged to call
+    // .update() on our behalf.
+  }
+
+  public void update() {
+    setSystemKeyboardPreference.setChecked(SystemIMESettings.isEnabledAsSystemKB(context));
+    setDefaultKeyboardPreference.setChecked(SystemIMESettings.isDefaultKB(context));
 
     // Update the language / keyboard count when we return to this menu from deeper levels.
     languagesPreference.setTitle(getInstalledLanguagesText());