+++ developer/js/lexical-model-compiler/build-trie.ts
@@ -33,16 +33,19 @@ export function createTrieDataStructure(sourceFiles: string[], searchTermToKey?:
  *
  * Format specification:
  *
- *  - the file is a UTF-8 encoded text file
- *  - new lines are either LF or CRLF
- *  - the file either consists of a comment or an entry
- *  - comment lines MUST start with the '#' character on the very first column
+ *  - the file is a UTF-8 encoded text file.
+ *  - new lines are either LF or CRLF.
+ *  - the file MAY start with the UTF-8 byte-order mark (BOM); that is, if the
+ *    first three bytes of the file are EF BB BF, these will be interepreted as
+ *    the BOM and will be ignored.
+ *  - the file either consists of a comment or an entry.
+ *  - comment lines MUST start with the '#' character on the very first column.
  *  - entries are one to three columns, separated by the (horizontal) tab
- *    character
+ *    character.
  *  - column 1 (REQUIRED): the wordform: can have any character except tab, CR,
  *    LF. Surrounding whitespace characters are trimmed.
  *  - column 2 (optional): the count: a non-negative integer specifying how many
- *    times this entry has appeared in the corpus. Blank means 'indeterminate'
+ *    times this entry has appeared in the corpus. Blank means 'indeterminate'.
  *  - column 3 (optional): comment: an informative comment, ignored by the tool.
  */
 export function parseWordList(contents: string): WordList {
@@ -54,6 +57,11 @@ export function parseWordList(contents: string): WordList {
 
   let result: WordList = [];
   for (let line of lines) {
+    // Remove the byte-order mark (BOM) from the beginning of the string.
+    // Because `contents` can be the concatenation of several files, we have to remove
+    // the BOM from every possible start of file -- i.e., beginning of every line.
+    line = line.replace(/^\uFEFF/, '');
+
     if (line.startsWith('#') || line === "") {
       continue; // skip comments and empty lines
     }