+++ ios/engine/KMEI/KeymanEngine/Classes/Manager.swift
@@ -308,6 +308,8 @@ public class Manager: NSObject, HTTPDownloadDelegate, UIGestureRecognizerDelegat
       _ = Manager.shared.registerLexicalModel(first_model)
     }
     
+    inputViewController.fixLayout()
+    
     return true
   }
 
@@ -507,8 +509,9 @@ public class Manager: NSObject, HTTPDownloadDelegate, UIGestureRecognizerDelegat
     guard let lm = removeLexicalModelFromUserList(userDefs: userData, at: index) else {
       return false
     }
-    
+
     removeLexicalModelFromLanguagePreference(userDefs: userData, lm)
+    inputViewController.deregisterLexicalModel(lm);
     // Set a new lexical model if deleting the current one
     let userLexicalModels = userData.userLexicalModels! //removeLexicalModelFromUserList fails above if this is not present
 
@@ -1537,7 +1540,8 @@ public class Manager: NSObject, HTTPDownloadDelegate, UIGestureRecognizerDelegat
   /// TextView/TextField to enable/disable the keyboard picker
   public func showKeyboardPicker(in viewController: UIViewController, shouldAddKeyboard: Bool) {
     hideKeyboard()
-    let vc = KeyboardPickerViewController()
+    let vc = shouldAddKeyboard ? KeyboardPickerViewController() :
+      KeyboardSwitcherViewController()
     let nc = UINavigationController(rootViewController: vc)
     nc.modalTransitionStyle = .coverVertical
     nc.modalPresentationStyle = .pageSheet